[CE-341]: K8S util job params fix (#10849)
* [CE-291]: Fixed K8S util job

* [CE-341]: setting time in k8s util obj

* [CE-341]: k8s resource unit fix

* [CE-341]: Account id in util

+++ 78-batch-processing/src/main/java/io/harness/batch/processing/billing/timeseries/data/K8sGranularUtilizationData.java
@@ -10,8 +10,8 @@ public class K8sGranularUtilizationData {
   private String instanceId;
   private String instanceType;
   private String settingId;
-  private long cpu;
-  private long memory;
+  private double cpu;
+  private double memory;
   private long endTimestamp;
   private long startTimestamp;
 }
